Move classes out of event dialog file.
