+++ app/lib/timetable/timetable_add_event/src/timetable_add_event_dialog_controller.dart
@@ -1,0 +1,114 @@
+import 'package:clock/clock.dart';
+import 'package:common_domain_models/common_domain_models.dart';
+import 'package:date/date.dart';
+import 'package:flutter/foundation.dart';
+import 'package:time/time.dart';
+
+import 'timetable_add_event_dialog_src.dart';
+
+class AddEventDialogController extends ChangeNotifier {
+  AddEventDialogController({required this.api, required this.isExam});
+  final EventDialogApi api;
+  final bool isExam;
+
+  String _title = '';
+
+  String get title => _title;
+  bool showEmptyTitleError = false;
+
+  set title(String value) {
+    showEmptyTitleError = value.isEmpty;
+    _title = value;
+    notifyListeners();
+  }
+
+  String _description = '';
+
+  String get description => _description;
+
+  set description(String value) {
+    _description = value;
+    notifyListeners();
+  }
+
+  CourseView? _course;
+
+  CourseView? get course => _course;
+
+  Future<void> selectCourse(CourseId courseId) async {
+    final c = await api.loadCourse(courseId);
+    _course = CourseView(id: courseId, name: c.name);
+    notifyListeners();
+  }
+
+  Date _date = Date.fromDateTime(clock.now());
+
+  Date get date => _date;
+
+  set date(Date value) {
+    _date = value;
+    notifyListeners();
+  }
+
+  Time _startTime = Time(hour: 11, minute: 00);
+
+  Time get startTime => _startTime;
+
+  set startTime(Time value) {
+    _startTime = value;
+    notifyListeners();
+  }
+
+  Time _endTime = Time(hour: 12, minute: 00);
+
+  Time get endTime => _endTime;
+
+  set endTime(Time value) {
+    _endTime = value;
+    notifyListeners();
+  }
+
+  Future<void> createEvent() async {
+    if (title.isEmpty) {
+      showEmptyTitleError = true;
+      notifyListeners();
+      return;
+    }
+    return api.createEvent(CreateEventCommand(
+      title: title,
+      description: description,
+      courseId: course!.id,
+      date: date,
+      startTime: startTime,
+      endTime: endTime,
+      location: location,
+      notifyCourseMembers: notifyCourseMembers,
+      eventType: isExam ? EventType.exam : EventType.event,
+    ));
+  }
+
+  bool _notifyCourseMembers = true;
+
+  bool get notifyCourseMembers => _notifyCourseMembers;
+
+  set notifyCourseMembers(bool value) {
+    _notifyCourseMembers = value;
+    notifyListeners();
+  }
+
+  String _location = '';
+
+  String get location => _location;
+
+  set location(String value) {
+    _location = value;
+    notifyListeners();
+  }
+}
+
+class CourseView {
+  final CourseId id;
+  final String name;
+
+  CourseView({required this.id, required this.name});
+}
